refactor: 필터링 상태 그룹 분리
